requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.9
python-decouple==3.8
django-filter==23.4
requests==2.31.0
gunicorn==21.2.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
django-storages[google]==1.14.2
Pillow==10.1.0
twilio==8.10.0